lab_env_requirements.txt
apturl==0.5.2
bcrypt==3.1.7
biopython==1.83
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.1
croniter==0.3.29
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==2.0.0
fasteners==0.14.1
fonttools==4.57.0
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-resources==6.4.5
Jinja2==2.10.1
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.7.5
monotonic==1.5
msgpack==0.6.2
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
prompt-toolkit==3.0.52
protobuf==3.6.1
psutil==5.5.1
pyarrow==17.0.0
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymol==2.3.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==3.1.4
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
pytz==2025.2
pyxdg==0.26
pyxlsb==1.0.10
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
ssh-import-id==5.10
systemd-python==234
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.14
xkit==0.0.0
xlrd==2.0.2
zipp==3.20.2
